move http-api.js to httpApi.ts
